QA loader: Reduce disk space of test files
Since we generate a fake file for replies and submissions, we
should limit the number of replies generated per iteration to 3
(previously it was 100). In addition, we need to use a very steep
file size distribution, or the total size of the files generated
is unwieldly (i.e. fills up the staging VM).